[REFACTOR] api views and url refactores to restfull api function based

+++ api/views.py
@@ -50,52 +50,40 @@ def getRoutes(request):
     return Response(routes)
 
 
-@api_view(['GET'])
+@api_view(['GET', 'POST'])
 def getNotes(request):
-    notes = Note.objects.all()
-    serializer = NoteSerializer(notes, many=True)
+    if request.method == 'GET':
+        notes = Note.objects.all()
+        serializer = NoteSerializer(notes, many=True)
+        return Response(serializer.data)
 
-    return Response(serializer.data)
-
-
-@api_view(['GET'])
-def getNote(request, pk):
-    notes = Note.objects.get(id=pk)
-    serializer = NoteSerializer(notes, many=False)
-
-    return Response(serializer.data)
-
-@api_view(['POST'])
-def createNote(request):
-    data = request.data
-    note = Note.objects.create(
-        body=data['body']
+    if request.method == 'POST':
+        data = request.data
+        note = Note.objects.create(
+            body=data['body']
         )
-    serializers = NoteSerializer(note, many=True)
-    
-    return Response(serializers.data)
+        serializers = NoteSerializer(note, many=True)
+        return Response(serializers.data)
 
 
-@api_view(['PUT'])
-def updateNote(request, pk):
-    """Update a note"""
+@api_view(['GET', 'POST', 'PUT', 'DELETE'])
+def getNote(request, pk):
+    if request.method == 'GET':
+        notes = Note.objects.get(id=pk)
+        serializer = NoteSerializer(notes, many=False)
+        return Response(serializer.data)
 
-    data = request.data
-    note = Note.objects.get(id=pk)
-    serializer = NoteSerializer(instance=note, data=data)
+    if request.method == 'PUT':
+        data = request.data
+        note = Note.objects.get(id=pk)
+        serializer = NoteSerializer(instance=note, data=data)
 
-    if serializer.is_valid():
-        serializer.save()
+        if serializer.is_valid():
+            serializer.save()
 
-    return Response(serializer.data)
+        return Response(serializer.data)
 
-
-@api_view(['DELETE'])
-def deleteNote(request, pk):
-    """Delete a note"""
-
-    note = Note.objects.get(id=pk)
-    note.delete()
-    
-    return Response('Note was deleted')
-
+    if request.method == 'DELETE':
+        note = Note.objects.get(id=pk)
+        note.delete()
+        return Response('Note was deleted')
